==> SCRIPTS/01-Orders-Data-Wrangling-&-Subsetting.ipynb <==
# Data Wrangling & Subsetting

## Content
#### 1. Importing libraries and data
#### 2. Data wrangling
#### 2.1. Dropping columns (and checking for missing values)
#### 2.2. Renaming columns
#### 2.3 Changing variable's data type
#### 2.4 Transposing data
#### 3. Data dictionary
#### 4. Changing data type of a column in df_ords
#### 5. Renaming unituitive column name in df_ords without overwriting dataframe
#### 6. Busiest hour for placing orders
#### 7. Meaning of value 4 in column department_id
#### 8. Creating subsets 'breakfast' and "dinner_party'
#### 9. Checking for count of rows in last dataframe
#### 10. Extracting information about user with user_id = 1 and doing basic stat
#### 11. Exporting data

# 1. Importing libraries and data

# Importing libraries
import pandas as pd
import numpy as np
import os

# Turning project folder path into a string
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Importing orders.csv, products.csv and departments.csv using os library
# Omitting index column if present as pandas automatically creates one
df_ords = pd.read_csv (os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)
df_prods = pd.read_csv (os.path.join (path, '02 Data', 'Original Data', 'products.csv'), index_col = False)
df_dep = pd.read_csv (os.path.join (path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# 2. Data wrangling

# Looking at first 5 rows of all 3 dataframes to get a sence of what wrangling procedures needs to be performed
df_ords.head()

df_prods.head()

df_dep.head()

## 2.1. Dropping columns (and checking for missing values)

# Overwritting df_ords with dataframe without column eval_set ("dropping" unnecessary column)
df_ords = df_ords.drop (columns = ['eval_set'])

# Checking for missing values in df_ords in column days_since_prior order
df_ords['days_since_prior_order'].value_counts(dropna = False)

## 2.2. Renaming columns

# Renaming column orders_dow in df_ords
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Checking column's new name
df_ords.head()

## 2.3 Changing variable's data type

# Changing data type of column order_id to string so that function describe() can skip this column
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Checking column's new data type
df_ords['order_id'].dtype

## 2.4 Transposing data

# Transposing df_dep and saving this transposed dataframe as a new one
df_dep_t = df_dep.T

# Checking how new dataframe looks
df_dep_t

# Removing 0 from header
df_dep_t.reset_index()            # Indexing all rows first

new_header = df_dep_t.iloc[0]     # Creating new header
df_dep_t_new = df_dep_t[1:]       # Copying all data from second row onward 
df_dep_t_new.columns = new_header # Adding new header

# Checking if all changes happened as desired
df_dep_t_new

# 3. Data dictionary

# Creating variable data dictionary
data_dict = df_dep_t_new.to_dict('index')

# Checking how it looks
data_dict

# 4. Changing data type of a column in df_ords

# Changing data type of numeric variable user_id in dataframe df_ords to string
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# 5. Renaming unituitive column name in df_ords without overwriting dataframe

# Creating "deep" copy of dataframe df_ords
df_ords_2 = df_ords.copy()

# Renaming column in new datframe df_ords_2 without overwriting original dataframe df_ords
df_ords_2.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

# Checking for new column name in df_ords_2
df_ords_2.head()

# Checking that df_ords wasn't overwritten
df_ords.head()

# 6. Busiest hour for placing orders

# Getting all frequencies for values in column order_hour_of_day
df_ords_2['order_hour_of_day'].value_counts()

# Answer: Busiest hour for placing orders is 10 A.M.

# 7. Meaning of value 4 in column department_id

# Using data dictionary to find what value 4 means
print(data_dict.get('4'))

# Answer: 4 means produce

# 8. Creating subsets 'breakfast' and "dinner_party'

# Using function loc to create subset 'breakfast'
df_breakfast = df_prods.loc[df_prods['department_id']==14]

# Checking the entire subset
df_breakfast

# This subset has 1116 rows

#### Creating subset dinner_party

# Using function isin to create subset with products: alcohol, deli, beverages and meat/seafood 
df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5, 7, 12, 20])]

# Checking entire subset
df_dinner_party

# This subset has 7650 rows

# 9. Checking for count of rows in last dataframe

# Count of rows in df_dinner_party
df_dinner_party.shape

# Answer: This dataframe has 7650 rows. (Already see in previous task.)

# 10. Extracting information about user with user_id = 1 and doing basic stat

# Getting all information for user with user_id = 1
df_user_1 = df_ords_2.loc[df_ords_2['user_id'] == '1']

# Showing all information found
df_user_1

#### Basic stat about user_1

# Getting basic descriptive statistic about user_1
df_user_1.describe()

# Getting count of values for last 3 columns
df_user_1['orders_day_of_week'].value_counts()

df_user_1['order_hour_of_day'].value_counts()

df_user_1['days_since_last_order'].value_counts()

# Days of the week (from Project brief):
# 0 = Saturday, 1 = Sunday, 2 = Monday, 3 = Tuesday, 4 = Wednesday, 5 = Thursday, 6 = Friday

# Answer:
# user_1 ordered 11 times, on days between Sunday and Wednesday, most frequently on Wednesday (4 times).
# Earliest at 7 A.M. and latest at 4 P.M. (16 in dataframe), most frequently, that is 3 times, at 7 A.M. and 8 A.M.
# Average hour of the day for orders was 10 A.M. and average day was Tuesday ( 2.64 is rounded up to 3 which is Tuesday).
# Minimum days that passed between 2 consecutive orders is 0 days, meaning orders with IDs 8 and 9 were made on the same day
# at 2 P.M. and 4 P.M. on Sunday.
# Maximum days passed between 2 consecutive orders were 30 days. Averaging with 19 days between 2 orders.


# 11. Exporting data

#### Trying to solve problem with code below not exporting columns order_id and user_id as strings, but again as int64 by adding quotes around nonnumeric values (quoting=2) with character ''df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled_1.csv'), index = False, quoting = 2, quotechar = '\'')
#### I excluded index column when exporting

# Exporting df_ords
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'), index = False)

# Exporting df_dep_t_new
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
