==> SCRIPTS/02-Orders-Data-Consistency-Checks.ipynb <==
# Data Consistency Checks

## Content
#### 1. Importing libraries and data
#### 2. Wrangling procedures that haven't been saved properly to orders_wrangled.csv
#### 3. Data consistency checks for df_prods
#### 3.1 Mixed type data
#### 3.2 Missing values
#### 3.3 Duplicates
#### 4. Data consistency check for df_ords
#### 4.1 Mixed type data
#### 4.2 Missing values
#### 4.3 Duplicates
#### 5. Exporting dataframes

## Comments about orders_wrangled.csv
#### In previous notebook (from task 4.4) I slightly changed code for exporting dataframe called there df_ords to orders_wrangled.csv so that entire index column is skipped and not exported orders_wrangled.csv. Unfortunately some steps of wrangling orders.csv are not properly saved/exported and I will need to change data types for 2 columns order_id and user_id again.

# 1. Importing libraries and data

# Importing libraries
import pandas as pd
import numpy as np
import os

# Project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Importing datasets products.csv i orders_wrangled.csv
df_prods = pd.read_csv (os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)
df_ords = pd.read_csv (os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# 2. Wrangling procedures that haven't been saved properly to orders_wrangled.csv

df_ords.head()

df_ords.info()

# Changing data types for columns order_id and user_id as it was done in Task 4.4
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')

# Checking data types again
df_ords.info()

# 3. Data consistency checks for df_prods

# Getting number of rows and columns before any changes are made
df_prods.shape

## 3.1 Mixed type data

# Check for mixed type columns
for col in df_prods.columns.tolist():
  weird_prods = (df_prods[[col]].applymap(type) != df_prods[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_prods[weird_prods]) > 0:
    print (col)

#### Function above printed column 'product_name' as column with mixed values.
#### It is possible missing values NaN are numeric instead of string.
#### I will procede with checking for missing values to see if this is the case.

## 3.2 Missing values

# Finding missing values in df_prods
df_prods.isnull().sum()

#### Indeed, missing values in column 'product_name' have numeric data type int64.

# Create subset of 16 missing values in column product_name
df_prods_nan = df_prods[df_prods['product_name'].isnull() == True]

df_prods_nan

# Create new dataframe from df_prods without missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

#### 16 missing values were removed successfully. At the same time those were the rows that made column 'product_name' column with mixed data dype.
#### Checking again if everything is consistent now. No column names printed means there aren't any columns with mixed data type.

for col in df_prods_clean.columns.tolist():
  weird_prods_clean = (df_prods_clean[[col]].applymap(type) != df_prods_clean[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_prods_clean[weird_prods_clean]) > 0:
    print (col)

## 3.3 Duplicates

# Create subset with full duplicates - rows that has duplicates in every column
df_prods_dups = df_prods_clean[df_prods_clean.duplicated()]

df_prods_dups

# Creating new dataframe without duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# 4. Data consistency check for df_ords

# Getting number of rows and columns before any changes are made
df_ords.shape

df_ords.head()

#### I notice that NaN in column 'days_since_prior_order' refers to N/A for first order from a new customer and this should be taken into consideration when dealing with such data.

df_ords.info()

df_ords.describe()

#### In all columns nothing looks out of the ordinary. For example, max for order_number is 100, but it is not unusual for someone to order 100 times from Instacart. 

## 4.1 Mixed data type

# Check for mixed type columns
for col in df_ords.columns.tolist():
  weird_ords = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird_ords]) > 0:
    print (col)

#### There are no columns with mixed data type

## 4.2 Missing values

# Finding missing values in df_ords
df_ords.isnull().sum()

#### There are 206209 missing valees in column days_since_prior_order.
#### As mentioned above if all this values are for the first order of a new customer, then they should be all left as is, because imputing this values with 0 would skew the results. 0 days means somebody ordered twice in one day.

# Create subset of all rows that have order_number = 1
df_ords_first_order = df_ords.loc[df_ords['order_number']==1]

df_ords_first_order.head()

df_ords_first_order.shape

#### All 206209 NaN values from column "days_since_prior_order" are values for first orders from a new costumer.
#### I propose all of this values to be left as they are. 

## 4.3 Duplicates

# Create subset with full duplicates - rows that has duplicates in every column
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

#### There are no duplicate values in this dataframe.

# 5. Exporting dataframes

#### For both dataframes, I am adding index_col = False so that exported files do not contain index column.

df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'), index = False)

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'), index = False)
