==> SCRIPTS/03-Combining-&-Exporting-Data-1.ipynb <==
# Combining & Exporting Data - Part 1

## Content
#### 1. Import libraries and data
#### 2. Data consistency check for df_ords_prior
#### 3. Combine df_ords and df_ords_prior
#### 4. Export df_merged_large

# 1. Import libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

# Create project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Import datasets orders_products_prior.csv and orders_checked.csv
df_ords_prior = pd.read_csv (os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)
df_ords = pd.read_csv (os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

# Check the output and shape for both df-s
df_ords_prior.head()

df_ords_prior.shape

df_ords.head()

df_ords.shape

# 2. Data consistency check for df_ords_prior

# Check for mixed type columns
for col in df_ords_prior.columns.tolist():
  weird = (df_ords_prior[[col]].applymap(type) != df_ords_prior[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords_prior[weird]) > 0:
    print (col)

#### No mixed data type columns found

# Find missing values in df_ords_prior
df_ords_prior.isnull().sum()

#### No missing values were found.

# Check for duplicates
# Create subset with full duplicates
df_ords_prior_dups = df_ords_prior[df_ords_prior.duplicated()]

# Check output
df_ords_prior_dups.head()

#### No duplicates were found.

# 3. Combine df_ords and df_ords_prior

# Merge df_ords and df_ords_prior with column 'order_id' as key
# Use default inner join and use merge flag
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check output
df_merged_large.head()

# Check if it's full merge
df_merged_large['_merge'].value_counts()

# Merge same 2 df-s with outer join to check if df_merged_large is indeed full merge
df_merged_large_out = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True, how = 'outer')

# Check output
df_merged_large_out.head()

# Check if it's full merge
df_merged_large_out['_merge'].value_counts()

# Create subset of rows with left_only flag
df_left = df_merged_large_out.loc[df_merged_large_out['_merge'] == 'left_only']

# Check output
df_left.head()

#### Inner merge is not full, but number 206209 of rows with left_only flags is very little compared to 32 millions rows and we will procede with df_merged_large

# Check shape of df_merged_large
df_merged_large.shape

# 4. Export df_merged_large

# Export df_merged_large as pikle format
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
