==> SCRIPTS/04-Combining-&-Exporting-Data-2.ipynb <==
# Combining & Exporting Data - Part 2

## Content
#### 1. Import libraries and data
#### 2. Merging data
#### 3. Exporting data

# 1. Import libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

# Create project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Import datasets products_checked.csv and orders_producs_combined.pkl
df_prods = pd.read_csv (os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)
df_op_comb = pd.read_pickle (os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))

# Check output and shape for both df-s
df_prods.head()

df_prods.shape

df_op_comb.head()

df_op_comb.shape

# 2. Merging data

#### Most suitable way to combine df_op_comb and df_prods would be through merging with inner join to avoid a lot of missing values and using column 'product_id' as key and using merge flag to see if merge was full
#### Column _merge in df_op_comb, representing flag from previous merge of another 2 df-s, will not be necessery in upcoming merge

# Drop column _merge from df_op_comb
df_op_comb_1 = df_op_comb.drop(columns = ['_merge'])

# Check output
df_op_comb_1.head()

# Merge df_op_comb_1 and df_prods with column 'product_id' as key and with merge flag
df_op_merged = df_op_comb_1.merge(df_prods, on = 'product_id', indicator = True)

# Check output
df_op_merged.head()

# Check shape
df_op_merged.shape

# Check if it's full merge
df_op_merged['_merge'].value_counts()

# Outer merge
df_op_merged_out = df_op_comb_1.merge(df_prods, on = 'product_id', indicator = True, how = 'outer')

# Check if it's full merge
df_op_merged_out['_merge'].value_counts()

#### df_op_merged does not represent full merge, but as it turns out all together 2029 + 11 = 2040 rows that are flaged with  left_only and right_only can be omitted (number is very little compared to 32 millions)
#### Merge flag is needed through merging process only and we can drop this column before exporting data.

# Drop merge flag
df_op_merged_1 = df_op_merged.drop(columns = ['_merge'])

# Check shape
df_op_merged_1.shape

# 3. Exporting data

#### Since df created through merging process has 32 million rows, data should be exported as pikle format

# Exporting df_op_merged_1
df_op_merged_1.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
