==> SCRIPTS/05-Deriving-New-Variables.ipynb <==
# Deriving New Variables

## Content
#### 1. Import libraries and data
#### 2. Create new column "price_range"
#### 3. Create new column "busiest day"
#### 4. Create new column "busiest days"
#### 5. Create new column 'busiest_period_of_day'
#### 6. Export data

# 1. Import libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

# Create project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Import dataset orders_products_merged.pkl
df_ords_prods_merged = pd.read_pickle (os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

df_ords_prods_merged.shape

df_ords_prods_merged.head()

# 2. Create new column "price_range"

#### Use loc() function to apply the 3 conditions, each in separate cell

# Define conditions

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range'] = 'High-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range'] = 'Mid-range product' 

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range'] = 'Low-range product'

# Count all values
df_ords_prods_merged['price_range'].value_counts(dropna = False)

# Check current shape and head of dataframe
df_ords_prods_merged.shape

df_ords_prods_merged.head()

# 3. Create new column "busiest day"

#### Create column with values implying how busy each day of the week was.
#### Run a for-loop for every row and assign desired value to one busiest day, one slowest day and to other 5 days asign regularly busy.

# Count all orders for each day
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

#### Busiest day is 0 = Saturday, least busy is 4 = Wednesday and the rest are regularly busy.

# Create an empty list first for for-loop

result = []

# for-loop for level of busyness of days of the week

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Create new column in dataframe
df_ords_prods_merged['busiest_day'] = result

# Count all values of new column
df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

# Check current shape and head of dataframe
df_ords_prods_merged.shape

df_ords_prods_merged.head()

# 4. Create new column "busiest days"

#### First drop column "busiest day" - it is no longer needed.
#### Run a for-loop for every row and assign desired value to 2 busiest days, 2 slowest days and to other 3 days asign regularly busy.
#### Busiest days are 0 = Saturday and 1 = Sunday, slowest days are 4 = Wednesday and 3 = Tuesday and the rest are Regularly busy

# Drop column 'busiest_day' and overwrite dataframe
df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['busiest_day'])

# Check if column was dropped successfully
df_ords_prods_merged.shape

df_ords_prods_merged.head()

# Create an empty list first for for-loop

result = []

# for-loop for level of busyness of days of the week

for value in df_ords_prods_merged["orders_day_of_week"]:
  if (value == 0 or value == 1):
    result.append("Busiest days")
  elif (value == 4 or value == 3):
    result.append("Slowest days")
  else:
    result.append("Regularly busy")

# Create new column in dataframe
df_ords_prods_merged['busiest_days'] = result

# Count all values of new column
df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

#### This numbers correspond to sums of orders for days that are of particular level of busyness, for example Busiest days = 11875462 = 6209632 + 5665830 which represent number of orders on days 0 = Saturday and 1 = Sunday. 

# Check current shape and head of dataframe
df_ords_prods_merged.shape

df_ords_prods_merged.head()

# 5. Create new column 'busiest_period_of_day'

# Count all orders for each hour of the day
df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

#### From this numbers we can see that most orders happened between hours 9 and 16 with each hour, having between around 2.4 M and 2.8 M orders. Least number of orders happened between hours 0 and 6 and at 23 with number of orders ranging from around 50 K to 400 K. At all other hours we have awerage number of order between about 600 K and 2.1 M. K=1000, M=1000000
#### I chose to group hours of days like this into 3 periods, because number of orders were similar in each group and I was looking for somewhat significant drop or rise between two hours in list above to determine when will period end or begin. For example, for hours 23 and 6 I noticed that for hours 21, 22 and 7 difference between number of orders is about 100 K, but between 23 and 6 it is 230 K and on the other side for 23, 6 and 0 difference in number of order is about 100K. 
#### This method actually produced 3 period of a day with equal lenght of 8 hours.

# Create an empty list first for for-loop

result = []

# for-loop for level of busyness of periods in a day

for value in df_ords_prods_merged["order_hour_of_day"]:
  if (9 <= value <= 16):
    result.append("Most orders")
  elif (0 <= value <= 6 or value == 23):
    result.append("Fewest orders")
  else:
    result.append("Awerage orders")

# Create new column in dataframe
df_ords_prods_merged['busiest_period_of_day'] = result

# Count all values of new column
df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

#### Numbers correspond to sums of number of order for each hour of the day that belongs to certain period. For example for label 'Fewest order' 1290198 = 218942 + 115786 + 69431 + 51317 + 53283 + 88057 + 290770 + 402612 which is a sum of number of order for hours 0,1, 2, 3, 4, 5, 6 and 23

# Check current shape and head of dataframe
df_ords_prods_merged.shape

df_ords_prods_merged.head()

# 6. Export data

# Export dataframe as pikle format
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_new_cols.pkl'))
