==> SCRIPTS/07-Wrangling-Customer-Data-&-Final-Merge.ipynb <==
# Intro to Data Visualization with Python - Part 1

## Content
#### 1. Import libraries and data
#### 2. Data wrangling of dataframe df_cust
#### 2.1. Rename columns
#### 2.2. Drop columns
#### 2.3. Columns data types
#### 2.4. Cleaning data
#### 3. Consistency check for dataframe df_cust
#### 3.1. Missing values
#### 3.2. Duplicates
#### 3.3. Mixed data types
#### 4. Additional data wrangling for dataframe df_ords_prods
#### 5. Combining 2 dataframes df_ords_prods and df_cust
#### 6. Export data

# 1. Import libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

# Create project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Import datasets orders_products_with_flags_new.pkl and customers.csv
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_with_flags_new.pkl'))
df_cust = pd.read_csv (os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

df_ords_prods.head()

df_ords_prods.shape

df_cust.head()

df_cust.shape

# 2. Data wrangling of dataframe df_cust

## 2.1. Rename columns

# Rename columns First Name, Surnam, STATE, Age, n_dependants, fam_status
df_cust.rename(columns = {'First Name' : 'first_name'}, inplace = True)
df_cust.rename(columns = {'Surnam' : 'surname'}, inplace = True)
df_cust.rename(columns = {'Gender' : 'gender'}, inplace = True)
df_cust.rename(columns = {'STATE' : 'state'}, inplace = True)
df_cust.rename(columns = {'Age' : 'age'}, inplace = True)
df_cust.rename(columns = {'n_dependants' : 'number_of_dependants'}, inplace = True)
df_cust.rename(columns = {'fam_status' : 'marital_status'}, inplace = True)

# Check output
df_cust.head()

## 2.2. Drop columns

# Show all unique values from column date_joined and their counts
df_cust['date_joined'].value_counts(dropna = False).sort_index()

#### This column is irrelevant for our future analysis and I'm dropping it to shorten dtaset before combining it with dataset orders_products_with_flags_new.

# Drop column date_joined and overwrite existing df_cust
df_cust = df_cust.drop (columns = ['date_joined'])

# Check output
df_cust.head()

## 2.3. Columns data types

# Check data types of all columns
df_cust.info()

#### All columns have the right data type. Numeric data types columns can be changed to use less memory and to optimize further steps.
#### It looks like only column 'first_name' has some missing values and I'll re-check this later.

# Check maximum value in all 4 numeric columns 'user_id' 'age', 'number_of_dependants' and 'income'
df_cust[['user_id', 'age', 'number_of_dependants', 'income']].max()

#### Based on these outputs, I will change data type for columns 'age' and 'number_of_dependants' to int8 and for 'user_id' and 'income' to int32

# Change data types in columns 'age', 'number_of_dependants' and 'income'
df_cust['age'] = df_cust['age'].astype('int8')
df_cust['number_of_dependants'] = df_cust['number_of_dependants'].astype('int8')
df_cust['user_id'] = df_cust['user_id'].astype('int32')
df_cust['income'] = df_cust['income'].astype('int32')

# Check all data types again and memory usage
df_cust.info()

## 2.4. Cleaning data

# Check if state names are correct
df_cust['state'].value_counts(dropna = False).sort_index()

#### All names are correct.

# 3. Consistency check for dataframe df_cust

## 3.1. Missing values

# Set Jupyter display option to show all rows
pd.options.display.max_rows = None

# Finding missing values in df_cust
df_cust.isnull().sum()

#### As suspected, column 'first_name' has 11259 missing values (NaN) which are of int64 data type. For now I'll leave them as they are.

## 3.2. Duplicates

# Create subset with full duplicates
df_cust_dups = df_cust[df_cust.duplicated()]

df_cust_dups

#### There are no duplicates.

## 3.3. Mixed data types

# Check for mixed type columns
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

#### I already concluded that this column has NaN values with data type int64 which is the reason why it was printed as column with "weird" type. For now I will leave them as they are.

# 4. Additional data wrangling for dataframe df_ords_prods

#### This dataframe has over 32 million of rows and 22 columns. I will check if some of the columns can be dropped or if some data types can be change to a data type that takes less memory. Data type for column 'user_id' needs to be change to int32 to match data type of column with same name in dataframe df_cust.

# Check data types for all columns
df_ords_prods.info()

# First change data type for column 'user_id'
df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('int32')

# Check output
df_ords_prods['user_id'].dtype

df_ords_prods.head(10)

#### Round to 2 decimal places all values in column 'mean_price' so that they resemble format of prices in stores.

df_ords_prods['mean_price'] = df_ords_prods['mean_price'].round(2)

df_ords_prods.head(10)

# Check maximum value for all columns with numeric data types (except 'user_id')
df_ords_prods[['order_id', 'order_number', 'orders_day_of_week', 'order_hour_of_day', 'days_since_prior_order', 'product_id', 'add_to_cart_order', 'reordered', 'aisle_id', 'department_id', 'prices', 'max_order', 'mean_price', 'median_days']].max()

# Change data type to int8
df_ords_prods['orders_day_of_week'] = df_ords_prods['orders_day_of_week'].astype('int8')
df_ords_prods['order_hour_of_day'] = df_ords_prods['order_hour_of_day'].astype('int8')
df_ords_prods['reordered'] = df_ords_prods['reordered'].astype('int8')
df_ords_prods['department_id'] = df_ords_prods['department_id'].astype('int8')

# Change data types to int16
df_ords_prods['order_number'] = df_ords_prods['order_number'].astype('int16')
df_ords_prods['add_to_cart_order'] = df_ords_prods['add_to_cart_order'].astype('int16')
df_ords_prods['aisle_id'] = df_ords_prods['aisle_id'].astype('int16')
df_ords_prods['max_order'] = df_ords_prods['max_order'].astype('int16')

# Change data types to int32
df_ords_prods['order_id'] = df_ords_prods['order_id'].astype('int32')
df_ords_prods['product_id'] = df_ords_prods['product_id'].astype('int32')

# Change data types to float32
df_ords_prods['days_since_prior_order'] = df_ords_prods['days_since_prior_order'].astype('float32')
df_ords_prods['prices'] = df_ords_prods['prices'].astype('float32')
df_ords_prods['mean_price'] = df_ords_prods['mean_price'].astype('float32')
df_ords_prods['median_days'] = df_ords_prods['median_days'].astype('float32')

# Check if all changes took place
df_ords_prods.info()

#### Changing all this data types resulted in lowering memory usage by about 2.4 GB, that is almost half of previous memory usage.
#### Column 'days_since_prior_order' needs to be of float data type because it contains NaN values.

# 5. Combining 2 dataframes df_ords_prods and df_cust

# Check shape for both dataframes
df_ords_prods.shape

df_cust.shape

#### First I'll merge 2 dataframes with outer merge and check how many rows will be flaged with merge flag 'only left' and 'only right' and if sum of them is insignificant to all the rows combined, then I'll merge them with inner merge and check if all rows will be flaged with merge flag 'both'.

# Merge df_ords_prods and df_cust with outer merge with key column 'user_id'
df_merge_out = df_ords_prods.merge(df_cust, on = 'user_id', indicator = True, how = 'outer')

# Check if it's full merge
df_merge_out['_merge'].value_counts()

#### This was full merge. No further action is needed.

# Check shape of merged dataframe
df_merge_out.shape

# Drop merge flag. It is no longer necessary.
df_merge_out = df_merge_out.drop(columns = ['_merge'])

# Check shape again and head()
df_merge_out.shape

df_merge_out.head()

# 6. Export data

# Export df_merge_out as pikle format
df_merge_out.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_all.pkl'))
