==> SCRIPTS/09-Deriving-New-Columns-&-Merging.ipynb <==
# Coding Etiquette & Excel Reporting (1)
# (Part 1 - Q 1-4)

## Content
#### 1. Import libraries and data
#### 2. Drop unnecessary columns and make new 'department' column
#### 3. Drop PII columns (Q2)
#### 4. Create new column 'region' (Q3)
#### 5. Create exclusion flag for low-activity customers (Q4)
#### 6. Export data

#### Note: After reading entire task and reviewing project brief I decided to drop all columns that are not useful for analysis and visualizations. I will also create new column 'department' with names of departments to have all necessary data in one place and to have more user-friendly visulizations later. (See section 2. for this code)

# 1. Import libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

# Create project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Import dataset orders_products_customers_with_dow_names.pkl
df_ords_prods_cust = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_customers_with_dow_names.pkl'))

# Import dataset departments_wrangled.csv ('index_col = False' didn't work here)
df_dep = pd.read_csv (os.path.join (path, '02 Data','Prepared Data', 'departments_wrangled.csv'), index_col = 0)

# Check head() and info() for df_ords_prods_cust
df_ords_prods_cust.head()

df_ords_prods_cust.info()

# Print all departments
df_dep

# Check all department_id
df_ords_prods_cust['department_id'].value_counts().sort_index()

# 2. Drop unnecessary columns and make new 'department' column

# Drop unnecessary columns
df_ords_prods_cust = df_ords_prods_cust.drop (columns = ['add_to_cart_order', 'reordered', 'aisle_id'])

# Check shape
df_ords_prods_cust.shape

# Make data dictionary out of df_dep to use for creation of new column
data_dict = df_dep.to_dict('index')

# Check dictionary
data_dict

# Create 'dictionary' from the output above
dictionary = {1:'frozen', 2:'other', 3:'bakery', 4:'produce',5:'alcohol', 6:'international', 7:'beverages', 8:'pets',
              9:'dry goods pasta', 10:'bulk', 11:'personal care', 12:'meat seafood', 13:'pantry', 14:'breakfast',
              15:'canned goods', 16:'dairy eggs', 17:'household', 18:'babies', 19:'snacks', 20:'deli', 21:'missing'}

# Make a new column 'department' using 'dictionary'
df_ords_prods_cust['department'] = df_ords_prods_cust['department_id'].map(dictionary)

# Check shape
df_ords_prods_cust.shape

# Check values in new column
df_ords_prods_cust['department'].value_counts().sort_index()

# Drop column 'department_id' and make new dataframe (to indicate that all data is in it)
df_all = df_ords_prods_cust.drop (columns = ['department_id'])

# Check shape
df_all.shape

# 3. Drop PII columns

#### Columns 'first_name' and 'surname' need to be dropped because they contain personally identifiable information (PII).

# Drop column 'first_name' and 'surname'
df_all = df_all.drop (columns = ['first_name', 'surname'])

# Check shape
df_all.shape

# 4. Create new column 'region'

#### Create regional segmentation of the data.
#### Information from this page was used to create values for this column: https://simple.wikipedia.org/wiki/List_of_regions_of_the_United_States

# Create 4 list of States for 4 regions
Northeast = ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']
Midwest = ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']
South = ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']
West = ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']

# Create new column 'region' with 4 values named after lists made above
df_all.loc[df_all['state'].isin(Northeast), 'region'] = 'Northeast'

df_all.loc[df_all['state'].isin(Midwest), 'region'] = 'Midwest'

df_all.loc[df_all['state'].isin(South), 'region'] = 'South'

df_all.loc[df_all['state'].isin(West), 'region'] = 'West'

# Check row counts per region
df_all['region'].value_counts(dropna = False)

#### Determine whether there’s a difference in spending habits between the different U.S. regions.

# Create crosstab
crosstab = pd.crosstab(df_all['region'], df_all['spending_flag'], dropna = False)

# Check output
crosstab

crosstab.to_clipboard()

#### Conclusion: Majority of Low spending users are from South and least of them come from Northeast region, by half compared to South. West and Midwest region have about the same proportion of users. The same goes for High spending users. 
#### This table can be found in Excel file Final report

# 5. Create exclusion flag for low-activity customers

#### Create an exclusion flag for low-activity customers (customers with less than 5 orders), exclude them from the data and export this sample.

# Create new column "activity-flag" based on their maximum order count
df_all.loc[df_all['max_order'] < 5, 'activity_flag'] = 'Low-activity customer'

df_all.loc[df_all['max_order'] >= 5, 'activity_flag'] = 'High-activity customer'

# Chech if new column was created
df_all.head()

# Check value counts
df_all['activity_flag'].value_counts(dropna = False)

# Create subset that excludes all low-activity customers
df_high_act = df_all[df_all['activity_flag'] == 'High-activity customer']

# Check if all low-activity customers were excluded correctly. If number is the same as in the above value counts, than yes.
df_high_act['activity_flag'].value_counts()

# 6. Export data

# Export separatelly df_all and df_high_act as pikle format
df_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'instacart_all_users.pkl'))

df_high_act.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'instacart_high_act_users.pkl'))
