==> SCRIPTS/10-Customer-Profiling-&-CrossTabs.ipynb <==
# Coding Etiquette & Excel Reporting (2)
# (Part 1 - Q 5)

## Content
#### 1. Import libraries and data
#### 2. Create a profiling variable
#### 2.1. Create new column 'age_group'
#### 2.2. Create new column 'family_size'
#### 2.3. Create new column 'income_class'
#### 2.4. Create new columns 'baby_flag' and 'pet_flag'
#### 2.5. Create new column 'period_of_day'
#### 2.6. Create new column 'income_class_and_age'
#### 2.7. Create one variable for customer profiling 'customer_profile'
#### 3. Drop unnecessary columns
#### 4. Export data

# 1. Import libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

# Create project folder path
path = r'C:\Users\Lara\Career Foundry Projects\21-09-2023 Instacart Basket Analysis'

# Import dataset instacart_high_act_users.pkl
df_all = pd.read_pickle (os.path.join (path, '02 Data','Prepared Data', 'instacart_high_act_users.pkl'))

# Check head() and info() of df_all
df_all.head()

df_all.info()

# 2. Create a customer profiling variable

#### Create a profiling variable based on age, income, certain goods in the “department_id” column, and number of dependents. You might also use the “orders_day_of_the_week” and “order_hour_of_day” columns if you can think of a way they would impact customer profiles. (Hint: As an example, try thinking of what characteristics would lead you to the profile “Single adult” or “Young parent.”)

## 2.1. Create new column 'age_group'

#### Age group intervals are made taking into consideration officially revised age groups made by WHO and UN in 2015. For people older than 60 I made just one group.


# Check min and max values in column 'age'
df_all['age'].min()

df_all['age'].max()

# Create new column 'age_group'
df_all.loc[df_all['age'] <= 25, 'age_group'] = 'Youth'

df_all.loc[(df_all['age'] > 25) & (df_all['age'] <= 44), 'age_group'] = 'Young adult'

df_all.loc[(df_all['age'] > 44) & (df_all['age'] <= 60), 'age_group'] = 'Middle-age adult'

df_all.loc[df_all['age'] > 60, 'age_group'] = 'Older adult'

# Check if new column was created
df_all.shape

# Check value counts
df_all['age_group'].value_counts(dropna = False)

#### Comment: The reason why age group Youth has low number of counts is because it describes only customers aged between 18 and 25 years with those years included, that is a span of 8 years. Other age groups span over about 15 - 20 years. Interenstingly, biggest customers are Older adults, that is people over 60 years old

# Create a custom made dictionary for new order/arrangement for age groups
arrange_age_gr = {'Youth' : 0, 'Young adult' : 1, 'Middle-age adult' : 2, 'Older adult' : 3} 

# Count all values in column 'age_group' and sort by custom made dictionary arrange_age_gr
df_all['age_group'].value_counts().sort_index(key = lambda x: x.map(arrange_age_gr))

# 2.2. Create new column 'family_size'

df_all['marital_status'].value_counts()

df_all['number_of_dependants'].value_counts().sort_index()

# Create crosstab to think about possible 'family_type' value names
crosstab_1 = pd.crosstab(df_all['marital_status'], df_all['number_of_dependants'], dropna = False)

crosstab_1

#### Comment: Dependant can be child, step child, sibling or parent or other close relative. It cannot be a spouse.
#### Crosstab above shows that single adults, or family of one, are people who's marrital status is either single or divorsed/widowed and they are the only ones without dependants. Family size for married people will be equal to number of dependants + 1, and for married people +2 ( because spouse is not a dependant!)

# Create new column 'family_size'
df_all.loc[df_all['number_of_dependants'] == 0, 'family_size'] = '1'
df_all.loc[((df_all['number_of_dependants'] == 1) & (df_all['marital_status'] == 'living with parents and siblings')), 'family_size'] = '2'
df_all.loc[(((df_all['number_of_dependants'] == 1) & (df_all['marital_status'] == 'married')) | ((df_all['number_of_dependants'] == 2) & (df_all['marital_status'] == 'living with parents and siblings'))) , 'family_size'] = '3'
df_all.loc[(((df_all['number_of_dependants'] == 2) & (df_all['marital_status'] == 'married')) | ((df_all['number_of_dependants'] == 3) & (df_all['marital_status'] == 'living with parents and siblings'))) , 'family_size'] = '4'
df_all.loc[((df_all['number_of_dependants'] == 3) & (df_all['marital_status'] == 'married')), 'family_size'] = '5'

# Check if new column was created
df_all.shape

# Check value counts
df_all['family_size'].value_counts(dropna = False).sort_index()

## 2.3. Create new column 'income_class'

# Check min and max of column 'income'
df_all['income'].min()

df_all['income'].max()

#### Calculations below were made with thought that 'income' column contains information about total family income.
#### Income classes were made taking into consideration conclusions from PEW reserch center about income classes in US for 2016. Middle income class family is defined as family of 3 (2 parents and 1 child) who has total family income between 2/3 and double of average US income, that is between 45,200 and 135,600 for year 2016. This should be further adjusted to various family sizes and place of living, but here I'll only adjust it to family size. All income values were rounded to nearest hundred. Link https://www.pewresearch.org/short-reads/2018/09/06/the-american-middle-class-is-stable-in-size-but-losing-ground-financially-to-upper-income-families/

# Create new column 'income_class'
df_all.loc[((df_all['family_size'] == '1') & (df_all['income'] < 26100)), 'income_class'] = 'Low income class'
df_all.loc[((df_all['family_size'] == '1') & ((df_all['income'] >= 26100) & (df_all['income'] <= 78200))), 'income_class'] = 'Middle income class'
df_all.loc[((df_all['family_size'] == '1') & (df_all['income'] > 78200)), 'income_class'] = 'Upper income class'

df_all.loc[((df_all['family_size'] == '2') & (df_all['income'] < 36900)), 'income_class'] = 'Low income class'
df_all.loc[((df_all['family_size'] == '2') & ((df_all['income'] >= 36900) & (df_all['income'] <= 110700))), 'income_class'] = 'Middle income class'
df_all.loc[((df_all['family_size'] == '2') & (df_all['income'] > 110700)), 'income_class'] = 'Upper income class'

df_all.loc[((df_all['family_size'] == '3') & (df_all['income'] < 45200)), 'income_class'] = 'Low income class'
df_all.loc[((df_all['family_size'] == '3') & ((df_all['income'] >= 45200) & (df_all['income'] <= 135600))), 'income_class'] = 'Middle income class'
df_all.loc[((df_all['family_size'] == '3') & (df_all['income'] > 135600)), 'income_class'] = 'Upper income class'

df_all.loc[((df_all['family_size'] == '4') & (df_all['income'] < 52200)), 'income_class'] = 'Low income class'
df_all.loc[((df_all['family_size'] == '4') & ((df_all['income'] >= 52200) & (df_all['income'] <= 156600))), 'income_class'] = 'Middle income class'
df_all.loc[((df_all['family_size'] == '4') & (df_all['income'] > 155600)), 'income_class'] = 'Upper income class'

df_all.loc[((df_all['family_size'] == '5') & (df_all['income'] < 58400)), 'income_class'] = 'Low income class'
df_all.loc[((df_all['family_size'] == '5') & ((df_all['income'] >= 58400) & (df_all['income'] <= 175100))), 'income_class'] = 'Middle income class'
df_all.loc[((df_all['family_size'] == '5') & (df_all['income'] > 175100)), 'income_class'] = 'Upper income class'

# Check if new column was created
df_all.shape

# Create a custom made dictionary for new order/arrangement for age groups
arrange_income_cls = {'Low income class' : 0, 'Middle income class' : 1, 'Upper income class' : 2} 

# Count all values and sort by arrange_income_cls
df_all['income_class'].value_counts(dropna = False).sort_index(key = lambda x: x.map(arrange_income_cls))

## 2.4. Create new columns 'baby_flag' and 'pet_flag'

# Create new column 'baby_flag'
df_all.loc[df_all['department'] == 'babies', 'baby_flag'] = 'Have baby'
df_all.loc[df_all['department'] != 'babies', 'baby_flag'] = 'Do not have baby'

# Check if new column was created
df_all.shape

# Check value counts
df_all['baby_flag'].value_counts(dropna = False)

# Group data by 'baby_flag' and count unique users
df_all.groupby('baby_flag')['user_id'].nunique()

crosstab_2 = pd.crosstab(df_all['marital_status'], df_all['baby_flag'], dropna = False)

crosstab_2

# Create new column 'pet_flag'
df_all.loc[df_all['department'] == 'pets', 'pet_flag'] = 'Have pet'
df_all.loc[df_all['department'] != 'pets', 'pet_flag'] = 'Do not have pet'

# Check if new column was created
df_all.shape

# Check value counts
df_all['pet_flag'].value_counts(dropna = False)

# Group data by 'pet_flag' and count unique users
df_all.groupby('pet_flag')['user_id'].nunique()

#### Comment: Even though 15% users buys baby products, from crosstab I can only conclude that people of all family structure buy baby stuff. For example: all "living with siblings and parents" are <=25 years and still buy baby stuff. Possibly siblings are babies. Also "single" users are from all age groups except "older adults" and they all don't have any dependants and still buy baby stuff. Total count of baby products is only 1.3% of all transactions (number of rows).  I will not use this information for customer profiling.
#### Similarly, I won't use 'pet_flag' either as about 6% users have pets and total count of pet realted item bought is only about 0.3% of all transactions (number of rows).

# Drop both columns
df_all = df_all.drop (columns = ['baby_flag', 'pet_flag'])

# Check shape
df_all.shape

## 2.5. Create new columns 'part_of_day'

#### I used partition of the day into 4 parts: morning, mid-day, late afternoon to early night and late night to have 4 periods of the day with 6 hours each, even though previous analyses showed majority of items were bought between 8 am and 4 pm (16).

# Create new column 'part_of_day'
df_all.loc[(df_all['order_hour_of_day'] >= 5) & (df_all['order_hour_of_day'] < 11), 'part_of_day'] = 'Morning'
df_all.loc[(df_all['order_hour_of_day'] >= 11) & (df_all['order_hour_of_day'] < 17 ), 'part_of_day'] = 'Mid-day'
df_all.loc[(df_all['order_hour_of_day'] >= 17) & (df_all['order_hour_of_day'] < 23), 'part_of_day'] = 'Late afternoon to early night'
df_all.loc[(df_all['order_hour_of_day'] >= 23) | (df_all['order_hour_of_day'] < 5), 'part_of_day'] = 'Late night'

# Check shape
df_all.shape

df_all['part_of_day'].value_counts(dropna = False)

## 2.6. Create new column 'income_class_and_age'

# Create crosstab for 'age_group' and 'income_class' to compare values from newly made column below
crosstab_3 = pd.crosstab(df_all['age_group'], df_all['income_class'], dropna = False)

crosstab_3

crosstab_3.to_clipboard()

#### Younger adult is Youth or Young adult (all ages <= 44)

df_all.loc[((df_all['age'] <= 44) & (df_all['income_class'] == 'Low income class')), 'income_class_and_age'] = 'Younger adult with low income'

df_all.loc[((df_all['age'] <= 44) & (df_all['income_class'] == 'Middle income class')), 'income_class_and_age'] = 'Younger adult with mid income'

df_all.loc[((df_all['age'] <= 44) & (df_all['income_class'] == 'Upper income class')), 'income_class_and_age'] = 'Younger adult with high income'

#### Mid-age adult

df_all.loc[((df_all['age_group'] == 'Middle-age adult') & (df_all['income_class'] == 'Low income class')), 'income_class_and_age'] = 'Mid-age adult with low income'

df_all.loc[((df_all['age_group'] == 'Middle-age adult') & (df_all['income_class'] == 'Middle income class')), 'income_class_and_age'] = 'Mid-age adult with mid income'

df_all.loc[((df_all['age_group'] == 'Middle-age adult') & (df_all['income_class'] == 'Upper income class')), 'income_class_and_age'] = 'Mid-age adult with high income'

#### Older adult

df_all.loc[((df_all['age_group'] == 'Older adult') & (df_all['income_class'] == 'Low income class')), 'income_class_and_age'] = 'Older adult with low income'

df_all.loc[((df_all['age_group'] == 'Older adult') & (df_all['income_class'] == 'Middle income class')), 'income_class_and_age'] = 'Older adult with mid income'

df_all.loc[((df_all['age_group'] == 'Older adult') & (df_all['income_class'] == 'Upper income class')), 'income_class_and_age'] = 'Older with high income'

# Check shape
df_all.shape

# Count all values
df_all['income_class_and_age'].value_counts(dropna = False)

# 2.7. Create one variable for customer profiling 'customer_profile'

#### Only users without dependats are single and divorced/widowed (family_size is 1) - they will be Single adults. All users who live with siblings and parents are under 25 but they all have 1, 2 or 3 dependants, so they'll be a Head of household under 26.
#### Married and age <=44 will be Young parent, from 45 to 59 mid age parent and for over 60 an older parent.

#### Single adults

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Younger adult with low income'), 'customer_profile'] = 'Single younger adult with low income'

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Mid-age adult with low income'), 'customer_profile'] = 'Single mid-age adult with low income'

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Older adult with low income'), 'customer_profile'] = 'Single older adult with low income'

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Younger adult with mid income'), 'customer_profile'] = 'Single younger adult with mid income'

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Mid-age adult with mid income'), 'customer_profile'] = 'Single mid-age adult with mid income'

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Older adult with mid income'), 'customer_profile'] = 'Single older adult with mid income'

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Younger adult with high income'), 'customer_profile'] = 'Single younger adult with high income'

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Mid-age adult with high income'), 'customer_profile'] = 'Single mid-age adult with high income'

df_all.loc[(df_all['family_size'] == '1')  & (df_all['income_class_and_age'] == 'Older adult with high income'), 'customer_profile'] = 'Single older adult with high income'

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

#### Head of household under 26

df_all.loc[(df_all['marital_status'] == 'living with parents and siblings')  & (df_all['income_class'] == 'Low income class'), 'customer_profile'] = 'Head of household under 26 with low income'

df_all.loc[(df_all['marital_status'] == 'living with parents and siblings')  & (df_all['income_class'] == 'Middle income class'), 'customer_profile'] = 'Head of household under 26 with mid income'

df_all.loc[(df_all['marital_status'] == 'living with parents and siblings')  & (df_all['income_class'] == 'Upper income class'), 'customer_profile'] = 'Head of household under 26 with high income'

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

#### Parent

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Younger adult with low income'), 'customer_profile'] = 'Younger parent with low income'

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Mid-age adult with low income'), 'customer_profile'] = 'Mid-age parent with low income'

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Older adult with low income'), 'customer_profile'] = 'Older parent with low income'

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Younger adult with mid income'), 'customer_profile'] = 'Younger parent with mid income'

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Mid-age adult with mid income'), 'customer_profile'] = 'Mid-age parent with mid income'

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Older adult with mid income'), 'customer_profile'] = 'Older parent with mid income'

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Younger adult with high income'), 'customer_profile'] = 'Younger parent with high income'

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Mid-age adult with high income'), 'customer_profile'] = 'Mid-age parent with high income'

df_all.loc[(df_all['marital_status'] == 'married')  & (df_all['income_class_and_age'] == 'Older adult with high income'), 'customer_profile'] = 'Older parent with high income'

# Check shape
df_all.shape

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

# Replese all NaN values with string 'Other customers'
df_all['customer_profile'] = df_all['customer_profile'].fillna('Other customers')

# Check value counts
df_all['customer_profile'].value_counts(dropna = False)

# 3. Drop unnecessary columns

# Drop unnecessary column
df_all = df_all.drop ('income_class_and_age', axis = 1)

# Check shape
df_all.shape

# 4. Export data

# Export df_all as pikle format
df_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'instacart_customer_profile.pkl'))
